requirements.txt
Django==1.8.7
django-admin-sortable2
django_compressor
django-debug-toolbar==1.4
django-haystack==2.4.1
django-flat-theme
django-smart-selects
django-mptt
djangorestframework==3.3.1
jinja2
mysqlclient
Pillow
sorl-thumbnail
